Split utils into smaller submodules
