+++ src/jutulgpt/human_in_the_loop/ui.py
@@ -18,7 +18,7 @@ from langgraph.types import interrupt
 
 from jutulgpt.logging.session import get_session_logger
 from jutulgpt.rag.utils import modify_doc_content
-from jutulgpt.utils import add_julia_context, remove_julia_context
+from jutulgpt.utils.code_parsing import add_julia_context, remove_julia_context
 
 from .interactions import (
     Action,
